backend/requirements.txt
django==4.2.27
djangorestframework==3.16.1
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
